Move public key repository to components-core
