+++ bcmc/src/main/java/com/adyen/checkout/bcmc/BcmcComponentProvider.kt
@@ -9,11 +9,11 @@ package com.adyen.checkout.bcmc
 
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.ViewModelStoreOwner
-import com.adyen.checkout.card.repository.PublicKeyRepository
 import com.adyen.checkout.components.PaymentComponentProvider
 import com.adyen.checkout.components.base.GenericPaymentMethodDelegate
 import com.adyen.checkout.components.base.lifecycle.viewModelFactory
 import com.adyen.checkout.components.model.paymentmethods.PaymentMethod
+import com.adyen.checkout.components.repository.PublicKeyRepository
 
 class BcmcComponentProvider : PaymentComponentProvider<BcmcComponent, BcmcConfiguration> {
     override operator fun get(
