+++ card/src/main/java/com/adyen/checkout/card/CardDelegate.kt
@@ -12,8 +12,8 @@ import com.adyen.checkout.card.api.model.Brand
 import com.adyen.checkout.card.data.CardType
 import com.adyen.checkout.card.data.DetectedCardType
 import com.adyen.checkout.card.data.ExpiryDate
-import com.adyen.checkout.card.repository.PublicKeyRepository
 import com.adyen.checkout.components.base.PaymentMethodDelegate
+import com.adyen.checkout.components.repository.PublicKeyRepository
 import com.adyen.checkout.components.ui.FieldState
 import kotlinx.coroutines.CoroutineScope
